test: fold OpenAI background abort suite
